Issue 631
Removed redundant cast warning.

git-svn-id: http://tungsten-replicator.googlecode.com/svn/trunk@1491 9f5e3340-5c9a-25cb-7c1c-c133a2e9e928

+++ replicator/src/java/com/continuent/tungsten/replicator/applier/MySQLApplier.java
@@ -235,7 +235,7 @@ public class MySQLApplier extends JdbcApplier
 
             if (columnSpec.isUnsigned() && isNegative)
             {
-                switch ((int) columnSpec.getLength())
+                switch (columnSpec.getLength())
                 {
                     case 1 :
                         valToInsert = TINYINT_MAX_VALUE + 1 + extractedVal;
